Verify input source type constructors
