+++ src/GraphQL.Analyzers/AnalyzerReleases.Unshipped.md
@@ -14,3 +14,4 @@ GQL006 | Usage | Warning | InputGraphTypeAnalyzer, [Documentation](https://graph
 GQL007 | Usage | Warning | InputGraphTypeAnalyzer, [Documentation](https://graphql-dotnet.github.io/docs/analyzers/GQL007_CanNotSetSourceField)
 GQL008 | Usage | Warning | FieldArgumentAnalyzer, [Documentation](https://graphql-dotnet.github.io/docs/analyzers/GQL008_DoNotUseObsoleteArgumentMethod)
 GQL009 | Usage | Error | AwaitableResolverAnalyzer, [Documentation](https://graphql-dotnet.github.io/docs/analyzers/GQL009_UseAsyncResolver)
+GQL010 | Usage | Error | InputGraphTypeAnalyzer, [Documentation](https://graphql-dotnet.github.io/docs/analyzers/GQL010_CanNotResolveInputObjectConstructor)
